requirements.txt
fastapi
langchain
langchain-text-splitters
langchain-community
langchain-ollama
langchain-google-genai
langchain-postgres
psycopg2-binary
pgvector
pypdf
jq
python-dotenv
uvicorn
python-multipart
python-frontmatter
